// src/app/commons/types/clasification.type.ts
export type CLASIFICATION_TYPE =
    | 'ingresosEconomicaCapitulos'
    | 'ingresosEconomicaArticulos'
    | 'ingresosEconomicaConceptos'
    | 'ingresosEconomicaEconomicos'
    | 'gastosOrganicaOrganicos'
    | 'gastosProgramaAreas'
    | 'gastosProgramaPoliticas'
    | 'gastosProgramaGrupos'
    | 'gastosProgramaProgramas'
    | 'gastosEconomicaCapitulos'
    | 'gastosEconomicaArticulos'
    | 'gastosEconomicaConceptos'
    | 'gastosEconomicaEconomicos';